move attach-bindings out of test as it doesn't contain test cases
