+++ actionview/test/ujs/src/attach-bindings.js
